+++ client/src/containers/App/App.jsx
@@ -2,20 +2,21 @@ import React, { Component } from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 import StoreProvider from "../../contexts/StoreProvider";
-import Register from "../../containers/Register/Register";
-import Login from "../../containers/Register/Login";
 import PrivateRoute from "../../routes/PrivateRoute";
-import Dashboard from "../../containers/Dashboard";
-import Course from "../../containers/Course/Course";
-import CourseSubject from "../../containers/Course/CourseSubject";
-import Navbar from "../../containers/Navbar/Navbar";
-import MatchableJoin from "../../containers/Matchable/MatchableJoin";
-import MatchableCreate from "../../containers/Matchable/MatchableCreate";
 import theme from "../../theme";
+
 import { menu_options } from '../../utils/constants';
 import { MainContainer } from './components';
-import CourseDisplay from "../Course/CourseDisplay";
-import CourseDetailV2 from "../Course/CourseDetailV2";
+
+import Course from "../Course/Course";
+import CourseDetail from "../Course/CourseDetail";
+import CourseSubject from "../Course/CourseSubject";
+import Dashboard from "../Dashboard";
+import MatchableJoin from "../Matchable/MatchableJoin";
+import MatchableCreate from "../Matchable/MatchableCreate";
+import Navbar from "../Navbar/Navbar";
+import Login from "../Register/Login";
+import Register from "../Register/Register";
 
 class App extends Component {
   render() {
@@ -30,7 +31,7 @@ class App extends Component {
             <MainContainer>
               <PrivateRoute exact path="/course" component={Course}/>
               <PrivateRoute exact path="/course/:subject" component={CourseSubject}/>
-              <PrivateRoute exact path="/course/:subject/:catalog_number" component={CourseDetailV2}/>
+              <PrivateRoute exact path="/course/:subject/:catalog_number" component={CourseDetail}/>
               <PrivateRoute exact path='/matchable/join' component={MatchableJoin} />
               <PrivateRoute exact path='/matchable/create' component={MatchableCreate} />
               <Switch>
